requirements.txt
# ============================================
# Revamp CoPilot ADK - MVP Requirements
# Python 3.10+ Required
# ============================================

# Core ADK Framework - this will pull in compatible versions of dependencies
google-adk>=1.18.0

# Google Cloud Services
google-cloud-bigquery>=3.14.0
python-dotenv>=1.0.0

# Note: The following are dependencies of google-adk and will be auto-installed
# with compatible versions. We list them here for documentation only.
# - google-cloud-aiplatform>=1.87.0
# - pydantic>=2.9.0
# - typing-extensions>=4.11.0
# - pydantic-settings

# Testing (for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development Tools (optional)
black>=23.0.0
ruff>=0.1.0

# Streamlit Frontend Requirements for ADK V1
streamlit>=1.28.0

